web: extract framed widget mount into widget_module.js (no behaviour change)
Move the framed-widget implementation out of widgets.js verbatim:
mountRouteIframeWidget (kind "iframe" + route) and mountModuleWidget
(kind "module": source fetch with its 25 s timeout, CSP/sandbox constants,
srcdoc assembly, parent fetch/resize bridge, disposer) now live in
web/modules/widget_module.js together with the WIDGET_FRAME_* constants and
the frame-height helpers. widgets.js keeps the page host, the mountTab
dispatcher, the card registry/mount tracking, disposeMountedWidgets and the
declarative renderer, and re-exports WIDGET_FRAME_* so existing imports keep
working. boundedNumber moves to widget_job.js, the shared policy module both
files already import. No CSP, sandbox, timeout, message-name or dispose-order
change.

Tests: _framed_widget_sources() also reads widget_module.js; negative pins
that previously covered the in-file framed code now run against that union;
the node source pin for the resize-bridge wiring reads both files. Docs:
ARCHITECTURE module tree/ownership map and DEVELOPMENT "Declarative widgets"
name the new module. Size ratchet: widgets.js lands at 1352 lines (band
1001-1500) and its BAND_PATHS entry stays None because the adjacent
transition rule treats a surviving band rationale as immutable.
